# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Chat APP API",
        "description": "API documentation for Chat Application - A social media chat application with posts, comments, likes, friendships, and real-time messaging",
        "contact": {
            "email": "support@chatapp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/friendships/suggested": {
            "get": {
                "tags": [
                    "Friendships"
                ],
                "summary": "Get suggested friends",
                "description": "Get list of suggested friends (users who were active yesterday and are not already friends)",
                "operationId": "766030c1cc5c8f432851ea39846bf896",
                "responses": {
                    "200": {
                        "description": "List of suggested friends",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Jane Doe"
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "jane@example.com"
                                            },
                                            "profile_image": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "last_seen_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/friendships": {
            "post": {
                "tags": [
                    "Friendships"
                ],
                "summary": "Send friend request",
                "description": "Send a friend request to another user",
                "operationId": "319f5e39ba4d597087010c4e8eb6b836",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID of the user to send friend request to",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Friend request sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "requester_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "addressee_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "pending"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or cannot add yourself as friend"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/friendships/{friendship}/accept": {
            "post": {
                "tags": [
                    "Friendships"
                ],
                "summary": "Accept friend request",
                "description": "Accept an incoming friend request",
                "operationId": "cb0a39f0c13f060dc01173bbbd7eab96",
                "parameters": [
                    {
                        "name": "friendship",
                        "in": "path",
                        "description": "Friendship ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friend request accepted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "requester_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "addressee_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "accepted"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - not the recipient of the friend request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/friendships/{friendship}/reject": {
            "post": {
                "tags": [
                    "Friendships"
                ],
                "summary": "Reject friend request",
                "description": "Reject an incoming friend request",
                "operationId": "8a256f3ec111cb7206e1f7bf4d520806",
                "parameters": [
                    {
                        "name": "friendship",
                        "in": "path",
                        "description": "Friendship ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friend request rejected successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "requester_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "addressee_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "rejected"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - not the recipient of the friend request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/friendships/requests": {
            "get": {
                "tags": [
                    "Friendships"
                ],
                "summary": "Get friend requests",
                "description": "Get all incoming and outgoing friend requests for the authenticated user",
                "operationId": "b8797ae22ae0edcac074ab4b0d7923ec",
                "responses": {
                    "200": {
                        "description": "List of friend requests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "outgoing": {
                                            "description": "Friend requests sent by the user",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "requester_id": {
                                                        "type": "integer"
                                                    },
                                                    "addressee_id": {
                                                        "type": "integer"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "addressee": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "incoming": {
                                            "description": "Friend requests received by the user",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "requester_id": {
                                                        "type": "integer"
                                                    },
                                                    "addressee_id": {
                                                        "type": "integer"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "requester": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentication using Laravel session-based authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}